Task 4: Фикс расположения моделей и схем
