--- src/models/__init__.py
@@ -1,9 +1,0 @@
-from db.db import Base
-
-from .entity import LoginHistory, User
-
-__all__ = [
-    "Base",
-    "User",
-    "LoginHistory",
-]
