+++ src/repositories/login_history.py
@@ -1,6 +1,6 @@
-from models import LoginHistory
+from models.login_history import LoginHistory
 from repositories.base import RepositoryDB
-from schemas.entity import LoginHistoryCreate, LoginHistoryInDB
+from schemas.login_history import LoginHistoryCreate, LoginHistoryInDB
 
 
 class LoginHistoryRepository(
