+++ src/schemas/user.py
@@ -1,0 +1,19 @@
+from uuid import UUID
+
+from pydantic import BaseModel
+
+
+class UserCreate(BaseModel):
+    login: str
+    password: str
+    first_name: str
+    last_name: str
+
+
+class UserInDB(BaseModel):
+    id: UUID
+    first_name: str
+    last_name: str
+
+    class Config:
+        from_attributes = True
